requirements.txt
Flask
flask-cors
splunk-handler
xgboost
numpy
requests
python-dotenv
